Fix -Wdeprecated-copy (Clang, but partly GCC)
The warning cannot be fixed for aggregates and it is suppressed in that case.
Sadly, GCC currently emits the warning at the use location and the fix only
works in Clang, I hope GCC will resolve the issue by adjusting diagnostic
location to the definition site, and no actions will be needed further.